requirements-dev.txt
pytest
pytest-flask
psycopg2-binary
python-dotenv